+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/excel/retail/returned/RetailReturnExportModel.java
@@ -83,13 +83,13 @@ public class RetailReturnExportModel extends BaseBo<RetailReturn> implements Exc
    * 商品数量
    */
   @ExcelProperty("商品数量")
-  private Integer totalNum;
+  private BigDecimal totalNum;
 
   /**
    * 赠品数量
    */
   @ExcelProperty("赠品数量")
-  private Integer giftNum;
+  private BigDecimal giftNum;
 
   /**
    * 操作时间
