+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/mappers/RetailOutSheetDetailMapper.java
@@ -2,6 +2,7 @@ package com.lframework.xingyun.sc.mappers;
 
 import com.lframework.starter.web.core.mapper.BaseMapper;
 import com.lframework.xingyun.sc.entity.RetailOutSheetDetail;
+import java.math.BigDecimal;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
@@ -30,7 +31,7 @@ public interface RetailOutSheetDetailMapper extends BaseMapper<RetailOutSheetDet
    * @param num
    * @return
    */
-  int addReturnNum(@Param("id") String id, @Param("num") Integer num);
+  int addReturnNum(@Param("id") String id, @Param("num") BigDecimal num);
 
   /**
    * 减少退货数量
@@ -39,5 +40,5 @@ public interface RetailOutSheetDetailMapper extends BaseMapper<RetailOutSheetDet
    * @param num
    * @return
    */
-  int subReturnNum(@Param("id") String id, @Param("num") Integer num);
+  int subReturnNum(@Param("id") String id, @Param("num") BigDecimal num);
 }
